+++ saleor/graphql/discount/sorters.py
@@ -1,8 +1,7 @@
 import graphene
 from django.db.models import Min, Q, QuerySet
 
-from ..channel.sorters import validate_channel_slug
-from ..core.types import ChannelSortInputObjectType
+from ..core.types import SortInputObjectType
 
 
 class SaleSortField(graphene.Enum):
@@ -21,16 +20,15 @@ class SaleSortField(graphene.Enum):
 
     @staticmethod
     def qs_with_value(queryset: QuerySet, channel_slug: str) -> QuerySet:
-        validate_channel_slug(channel_slug)
         return queryset.annotate(
             value=Min(
                 "channel_listings__discount_value",
-                filter=Q(channel_listings__channel__slug=channel_slug),
+                filter=Q(channel_listings__channel__slug=str(channel_slug)),
             )
         )
 
 
-class SaleSortingInput(ChannelSortInputObjectType):
+class SaleSortingInput(SortInputObjectType):
     class Meta:
         sort_enum = SaleSortField
         type_name = "sales"
@@ -54,26 +52,24 @@ class VoucherSortField(graphene.Enum):
 
     @staticmethod
     def qs_with_minimum_spent_amount(queryset: QuerySet, channel_slug: str) -> QuerySet:
-        validate_channel_slug(channel_slug)
         return queryset.annotate(
             min_spent_amount=Min(
                 "channel_listings__min_spent_amount",
-                filter=Q(channel_listings__channel__slug=channel_slug),
+                filter=Q(channel_listings__channel__slug=str(channel_slug)),
             )
         )
 
     @staticmethod
     def qs_with_value(queryset: QuerySet, channel_slug: str) -> QuerySet:
-        validate_channel_slug(channel_slug)
         return queryset.annotate(
             discount_value=Min(
                 "channel_listings__discount_value",
-                filter=Q(channel_listings__channel__slug=channel_slug),
+                filter=Q(channel_listings__channel__slug=str(channel_slug)),
             )
         )
 
 
-class VoucherSortingInput(ChannelSortInputObjectType):
+class VoucherSortingInput(SortInputObjectType):
     class Meta:
         sort_enum = VoucherSortField
         type_name = "vouchers"