+++ saleor/product/utils/__init__.py
@@ -62,7 +62,7 @@ def delete_categories(categories_ids: List[str], manager):
 
 def collect_categories_tree_products(category: "Category") -> "QuerySet[Product]":
     """Collect products from all levels in category tree."""
-    products = category.products.prefetched_for_webhook(single_object=False)
+    products = category.products.prefetched_for_webhook(single_object=False)  # type: ignore
     descendants = category.get_descendants()
     for descendant in descendants:
         products = products | descendant.products.all()